Remove 'check' function, use cache mutex instead of withoutOverlapping() to solve the problem Tested OK

+++ src/QueueManager.php
@@ -2,6 +2,7 @@
 
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Cache\Repository;
 use Illuminate\Mail\Mailer;
 use Illuminate\Support\Collection;
 use Log;
@@ -11,15 +12,15 @@ use NZTim\Queue\QueuedJob\QueuedJobRepository;
 class QueueManager
 {
     protected $attempts;
+    protected $cache;
     protected $queuedJobRepo;
     protected $mutexHandler;
-    protected $laravelMailer;
 
-    public function __construct(QueuedJobRepository $queuedJobRepo, Mailer $laravelMailer)
+    public function __construct(QueuedJobRepository $queuedJobRepo, Repository $cache)
     {
         $this->attempts = env('QUEUEMGR_ATTEMPTS', 5);
         $this->queuedJobRepo = $queuedJobRepo;
-        $this->laravelMailer = $laravelMailer;
+        $this->cache = $cache;
     }
 
     // Facade commands --------------------------------------------------------
@@ -34,7 +35,19 @@ class QueueManager
 
     public function process()
     {
+        $cacheKey = 'nztim-queuemgr-lock';
+        if($this->cache->has($cacheKey)) {
+            Log::warning("QueueMgr triggered but process already running");
+            return;
+        }
+        $this->cache->put($cacheKey, true, 60);
         $this->queuedJobRepo->purgeDeleted();
+        $this->executeJobs();
+        $this->cache->forget($cacheKey);
+    }
+
+    protected function executeJobs()
+    {
         $queue = $this->queuedJobRepo->allOutstanding();
         foreach($queue as $item) {
             /** @var QueuedJob $item */
@@ -76,36 +89,5 @@ class QueueManager
     public function clearFailed()
     {
         $this->queuedJobRepo->clearFailed();
-    }
-
-    public function check()
-    {
-        $queue = $this->queuedJobRepo->allOutstanding();
-        $maxAgeInHours = env('QUEUEMGR_MAX_AGE', 1);
-        $maxAge = Carbon::now()->subHours($maxAgeInHours);
-        $exceeded = false;
-        foreach ($queue as $job) {
-            /** @var QueuedJob $job */
-            if (!$job->failed() && $job->created_at < $maxAge) {
-                $exceeded = true;
-            }
-        }
-        if (!$exceeded)  {
-            return;
-        }
-        $message = $this->message($maxAgeInHours);
-        Log::error($message);
-        $email = env('QUEUEMGR_EMAIL', null);
-        if (!empty($email) && filter_var($email, FILTER_VALIDATE_EMAIL)) {
-            $this->laravelMailer->raw($message, function ($message) use ($email) {
-                $message->to($email)->subject('QueueMgr job queue failure on ' . url('/'));
-            });
-        }
-    }
-
-    protected function message($maxAgeInHours)
-    {
-        $url = url('/');
-        return "The QueueMgr jobs queue on {$url} has exceeded the maximum age ({$maxAgeInHours} hours). There may be a server problem, such as the mutex file not being cleared.";
     }
 }
